♻️ factor out new 'mint in store' function
